refactor: 네이밍 수정 및 폴더구조 수정
