+++ src/Router.tsx
@@ -16,8 +16,8 @@ import CustomSizePage from './page/Custom/Common/CustomSizePage';
 import CustomImgLayout from './components/Custom/NoDesign/Img/CustomImgLayout';
 import CustomRequestLayout from './components/Custom/NoDesign/Request/CustomRequestLayout';
 import CustomThemePage from './page/Custom/HaveDesign/CustomThemeLayout';
-import AdditionalRequestPage from './page/Custom/HaveDesign/AdditionalRequestLayout';
-import PricePage from './page/Custom/PriceLayout';
+import AdditionalRequestPage from './components/Custom/HaveDesign/AdditionalRequestLayout';
+import PricePage from './components/Custom/Common/PriceLayout';
 import ReceiptPage from './page/Custom/HaveDesign/ReceiptLayout';
 import StylingColorPage from './page/Custom/HaveDesign/StylingColorLayout';
 import CustomReferencePage from './page/Custom/HaveDesign/CustomReferenceLayout';
