--- src/components/Custom/HaveDesign/AdditionalRequestFooter.tsx
@@ -1,40 +1,0 @@
-import { useNavigate } from 'react-router-dom';
-import { styled } from 'styled-components';
-
-const AdditionalRequestFooter = () => {
-  const navigate = useNavigate();
-
-  const handleClickButton = () => {
-    navigate('/');
-  };
-
-  return (
-    <St.CustomFooter>
-      <St.FooterButton type='button' onClick={handleClickButton}>
-        다음
-      </St.FooterButton>
-    </St.CustomFooter>
-  );
-};
-
-const St = {
-  CustomFooter: styled.footer`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    height: 7rem;
-    background-color: ${({ theme }) => theme.colors.gray9};
-  `,
-
-  FooterButton: styled.button`
-    width: 100%;
-    height: 100%;
-    color: ${({ theme }) => theme.colors.pink5};
-    ${({ theme }) => theme.fonts.title_semibold_18};
-  `,
-};
-
-export default AdditionalRequestFooter;
