+++ src/components/Custom/HaveDesign/CustomTheme/CustomThemeLayout.tsx
@@ -1,12 +1,13 @@
 import { useState } from 'react';
-import CustomTheme from '../../../components/Custom/HaveDesign/CustomTheme';
-import PageLayout from '../../../components/PageLayout';
-import Header from '../../../components/Header';
-import CancelBtn from '../../../common/Header/CancelBtn';
-import CustomSizeEscapeModal from '../../../common/Modal/EscapeModal/CustomSizeEscapeModal';
-import ProgressBar from '../../../common/ProgressBar';
-import { IcBackDark } from '../../../assets/icon';
-import HaveDesignFooter from '../../../components/Custom/HaveDesign/HaveDesignFooter';
+import { IcBackDark } from '../../../../assets/icon';
+import CancelBtn from '../../../../common/Header/CancelBtn';
+import CustomSizeEscapeModal from '../../../../common/Modal/EscapeModal/CustomSizeEscapeModal';
+import ProgressBar from '../../../../common/ProgressBar';
+import Header from '../../../Header';
+import PageLayout from '../../../PageLayout';
+import HaveDesignFooter from '../HaveDesignFooter';
+import CustomTheme from './CustomTheme';
+
 interface CustomThemeLayoutProps {
   setStep: React.Dispatch<React.SetStateAction<number>>;
   setName: React.Dispatch<React.SetStateAction<string>>;
