+++ src/components/Custom/HaveDesign/Receipt/ReceiptDetail.tsx
