+++ src/components/Custom/HaveDesign/Receipt/ReceiptFooter.tsx
