+++ src/components/Custom/HaveDesign/Reference/CustomReferenceLayout.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
-import CancelBtn from '../../../common/Header/CancelBtn';
-import CustomReference from '../../../components/Custom/HaveDesign/CustomReference';
-import Header from '../../../components/Header';
-import PageLayout from '../../../components/PageLayout';
-import ProgressBar from '../../../common/ProgressBar';
-import CustomSizeEscapeModal from '../../../common/Modal/EscapeModal/CustomSizeEscapeModal';
-import { IcBackDark } from '../../../assets/icon';
-import HaveDesignFooter from '../../../components/Custom/HaveDesign/HaveDesignFooter';
+import { IcBackDark } from '../../../../assets/icon';
+import CancelBtn from '../../../../common/Header/CancelBtn';
+import CustomSizeEscapeModal from '../../../../common/Modal/EscapeModal/CustomSizeEscapeModal';
+import ProgressBar from '../../../../common/ProgressBar';
+import Header from '../../../Header';
+import PageLayout from '../../../PageLayout';
+import HaveDesignFooter from '../HaveDesignFooter';
+import CustomReference from './CustomReference';
 
 interface CustomReferenceLayoutProps {
   setStep: React.Dispatch<React.SetStateAction<number>>;
